LayoutTests/fast/encoding/char-decoding-expected.txt

This tests decoding characters in various character sets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS decode('UTF-8', '%E2%88%9A') is 'U+221A'
PASS decode('gb2312', '%A3%A0') is 'U+3000'
PASS decode('gb_2312-80', '%A3%A0') is 'U+3000'
PASS decode('chinese', '%A3%A0') is 'U+3000'
PASS decode('gbk', '%A3%A0') is 'U+3000'
PASS decode('gb18030', '%A3%A0') is 'U+3000'
PASS decode('EUC-CN', '%A3%A0') is 'U+3000'
PASS decode('Shift_JIS', '%82%d0') is 'U+3072'
PASS decode('shift-jis', '%82%d0') is 'U+3072'
PASS decode('korean', '%A2%E6') is 'U+20AC'
PASS decode('korean', '%A1%A4') is 'U+00B7'
PASS decode('korean', '%A1%A9') is 'U+00AD'
PASS decode('korean', '%A1%AA') is 'U+2015'
PASS decode('korean', '%A1%AD') is 'U+223C'
PASS decode('korean', '%A2%A6') is 'U+FF5E'
PASS decode('korean', '%A2%C1') is 'U+2299'
PASS decode('korean', '%1A') is 'U+001A'
PASS decode('korean', '%1C') is 'U+001C'
PASS decode('korean', '%8F%A1') is 'U+B8EA'
PASS decode('korean', '%B4%D3') is 'U+B2D2'
PASS decode('korean', '%A2%41') is 'U+C910'
PASS decode('EUC-KR', '%A2%E6') is 'U+20AC'
PASS decode('EUC-KR', '%A1%A4') is 'U+00B7'
PASS decode('EUC-KR', '%A1%A9') is 'U+00AD'
PASS decode('EUC-KR', '%A1%AA') is 'U+2015'
PASS decode('EUC-KR', '%A1%AD') is 'U+223C'
PASS decode('EUC-KR', '%A2%A6') is 'U+FF5E'
PASS decode('EUC-KR', '%A2%C1') is 'U+2299'
PASS decode('EUC-KR', '%1A') is 'U+001A'
PASS decode('EUC-KR', '%1C') is 'U+001C'
PASS decode('EUC-KR', '%8F%A1') is 'U+B8EA'
PASS decode('EUC-KR', '%B4%D3') is 'U+B2D2'
PASS decode('EUC-KR', '%A2%41') is 'U+C910'
PASS decode('windows-949', '%A2%E6') is 'U+20AC'
PASS decode('windows-949', '%A1%A4') is 'U+00B7'
PASS decode('windows-949', '%A1%A9') is 'U+00AD'
PASS decode('windows-949', '%A1%AA') is 'U+2015'
PASS decode('windows-949', '%A1%AD') is 'U+223C'
PASS decode('windows-949', '%A2%A6') is 'U+FF5E'
PASS decode('windows-949', '%A2%C1') is 'U+2299'
PASS decode('windows-949', '%1A') is 'U+001A'
PASS decode('windows-949', '%1C') is 'U+001C'
PASS decode('windows-949', '%8F%A1') is 'U+B8EA'
PASS decode('windows-949', '%B4%D3') is 'U+B2D2'
PASS decode('windows-949', '%A2%41') is 'U+C910'
PASS decode('x-windows-949', '%A2%E6') is 'U+20AC'
PASS decode('x-windows-949', '%A1%A4') is 'U+00B7'
PASS decode('x-windows-949', '%A1%A9') is 'U+00AD'
PASS decode('x-windows-949', '%A1%AA') is 'U+2015'
PASS decode('x-windows-949', '%A1%AD') is 'U+223C'
PASS decode('x-windows-949', '%A2%A6') is 'U+FF5E'
PASS decode('x-windows-949', '%A2%C1') is 'U+2299'
PASS decode('x-windows-949', '%1A') is 'U+001A'
PASS decode('x-windows-949', '%1C') is 'U+001C'
PASS decode('x-windows-949', '%8F%A1') is 'U+B8EA'
PASS decode('x-windows-949', '%B4%D3') is 'U+B2D2'
PASS decode('x-windows-949', '%A2%41') is 'U+C910'
PASS decode('x-uhc', '%A2%E6') is 'U+20AC'
PASS decode('x-uhc', '%A1%A4') is 'U+00B7'
PASS decode('x-uhc', '%A1%A9') is 'U+00AD'
PASS decode('x-uhc', '%A1%AA') is 'U+2015'
PASS decode('x-uhc', '%A1%AD') is 'U+223C'
PASS decode('x-uhc', '%A2%A6') is 'U+FF5E'
PASS decode('x-uhc', '%A2%C1') is 'U+2299'
PASS decode('x-uhc', '%1A') is 'U+001A'
PASS decode('x-uhc', '%1C') is 'U+001C'
PASS decode('x-uhc', '%8F%A1') is 'U+B8EA'
PASS decode('x-uhc', '%B4%D3') is 'U+B2D2'
PASS decode('x-uhc', '%A2%41') is 'U+C910'
PASS decode('iso-ir-149', '%A2%E6') is 'U+20AC'
PASS decode('iso-ir-149', '%A1%A4') is 'U+00B7'
PASS decode('iso-ir-149', '%A1%A9') is 'U+00AD'
PASS decode('iso-ir-149', '%A1%AA') is 'U+2015'
PASS decode('iso-ir-149', '%A1%AD') is 'U+223C'
PASS decode('iso-ir-149', '%A2%A6') is 'U+FF5E'
PASS decode('iso-ir-149', '%A2%C1') is 'U+2299'
PASS decode('iso-ir-149', '%1A') is 'U+001A'
PASS decode('iso-ir-149', '%1C') is 'U+001C'
PASS decode('iso-ir-149', '%8F%A1') is 'U+B8EA'
PASS decode('iso-ir-149', '%B4%D3') is 'U+B2D2'
PASS decode('iso-ir-149', '%A2%41') is 'U+C910'
PASS decode('KS_C_5601-1987', '%A2%E6') is 'U+20AC'
PASS decode('KS_C_5601-1987', '%A1%A4') is 'U+00B7'
PASS decode('KS_C_5601-1987', '%A1%A9') is 'U+00AD'
PASS decode('KS_C_5601-1987', '%A1%AA') is 'U+2015'
PASS decode('KS_C_5601-1987', '%A1%AD') is 'U+223C'
PASS decode('KS_C_5601-1987', '%A2%A6') is 'U+FF5E'
PASS decode('KS_C_5601-1987', '%A2%C1') is 'U+2299'
PASS decode('KS_C_5601-1987', '%1A') is 'U+001A'
PASS decode('KS_C_5601-1987', '%1C') is 'U+001C'
PASS decode('KS_C_5601-1987', '%8F%A1') is 'U+B8EA'
PASS decode('KS_C_5601-1987', '%B4%D3') is 'U+B2D2'
PASS decode('KS_C_5601-1987', '%A2%41') is 'U+C910'
PASS decode('KS_C_5601-1989', '%A2%E6') is 'U+20AC'
PASS decode('KS_C_5601-1989', '%A1%A4') is 'U+00B7'
PASS decode('KS_C_5601-1989', '%A1%A9') is 'U+00AD'
PASS decode('KS_C_5601-1989', '%A1%AA') is 'U+2015'
PASS decode('KS_C_5601-1989', '%A1%AD') is 'U+223C'
PASS decode('KS_C_5601-1989', '%A2%A6') is 'U+FF5E'
PASS decode('KS_C_5601-1989', '%A2%C1') is 'U+2299'
PASS decode('KS_C_5601-1989', '%1A') is 'U+001A'
PASS decode('KS_C_5601-1989', '%1C') is 'U+001C'
PASS decode('KS_C_5601-1989', '%8F%A1') is 'U+B8EA'
PASS decode('KS_C_5601-1989', '%B4%D3') is 'U+B2D2'
PASS decode('KS_C_5601-1989', '%A2%41') is 'U+C910'
PASS decode('KSC5601', '%A2%E6') is 'U+20AC'
PASS decode('KSC5601', '%A1%A4') is 'U+00B7'
PASS decode('KSC5601', '%A1%A9') is 'U+00AD'
PASS decode('KSC5601', '%A1%AA') is 'U+2015'
PASS decode('KSC5601', '%A1%AD') is 'U+223C'
PASS decode('KSC5601', '%A2%A6') is 'U+FF5E'
PASS decode('KSC5601', '%A2%C1') is 'U+2299'
PASS decode('KSC5601', '%1A') is 'U+001A'
PASS decode('KSC5601', '%1C') is 'U+001C'
PASS decode('KSC5601', '%8F%A1') is 'U+B8EA'
PASS decode('KSC5601', '%B4%D3') is 'U+B2D2'
PASS decode('KSC5601', '%A2%41') is 'U+C910'
PASS decode('KSC_5601', '%A2%E6') is 'U+20AC'
PASS decode('KSC_5601', '%A1%A4') is 'U+00B7'
PASS decode('KSC_5601', '%A1%A9') is 'U+00AD'
PASS decode('KSC_5601', '%A1%AA') is 'U+2015'
PASS decode('KSC_5601', '%A1%AD') is 'U+223C'
PASS decode('KSC_5601', '%A2%A6') is 'U+FF5E'
PASS decode('KSC_5601', '%A2%C1') is 'U+2299'
PASS decode('KSC_5601', '%1A') is 'U+001A'
PASS decode('KSC_5601', '%1C') is 'U+001C'
PASS decode('KSC_5601', '%8F%A1') is 'U+B8EA'
PASS decode('KSC_5601', '%B4%D3') is 'U+B2D2'
PASS decode('KSC_5601', '%A2%41') is 'U+C910'
PASS decode('iso-8859-9', '%80') is 'U+20AC'
PASS decode('iso-8859-9', '%9F') is 'U+0178'
PASS decode('iso-8859-9', '%FD') is 'U+0131'
PASS decode('latin5', '%80') is 'U+20AC'
PASS decode('latin5', '%9F') is 'U+0178'
PASS decode('latin5', '%FD') is 'U+0131'
PASS decode('windows-1254', '%80') is 'U+20AC'
PASS decode('windows-1254', '%9F') is 'U+0178'
PASS decode('windows-1254', '%FD') is 'U+0131'
PASS decode('ISO-8859-13', '%A1') is 'U+201D'
PASS decode('ISO-8859-13', '%A5') is 'U+201E'
PASS decode('ISO-8859-13', '%B4') is 'U+201C'
PASS decode('ISO-8859-13', '%FF') is 'U+2019'
PASS decode('windows-1257', '%80') is 'U+20AC'
PASS decode('windows-1257', '%B4') is 'U+00B4'
PASS decode('windows-1257', '%FF') is 'U+02D9'
PASS decode('iso-8859-7', '%A1') is 'U+2018'
PASS decode('iso-8859-7', '%B5') is 'U+0385'
PASS decode('iso-8859-7', '%B6') is 'U+0386'
PASS decode('windows-1253', '%80') is 'U+20AC'
PASS decode('windows-1253', '%A1') is 'U+0385'
PASS decode('windows-1253', '%B5') is 'U+00B5'
PASS decode('windows-1253', '%B6') is 'U+00B6'
PASS decode('KOI8-R', '%A4') is 'U+2553'
PASS decode('KOI8-R', '%AD') is 'U+255C'
PASS decode('KOI8-U', '%A4') is 'U+0454'
PASS decode('KOI8-U', '%AD') is 'U+0491'
PASS decode('TIS-620', '%80') is 'U+20AC'
PASS decode('TIS-620', '%96') is 'U+2013'
PASS decode('TIS-620', '%A0') is 'U+00A0'
PASS decode('TIS-620', '%A1') is 'U+0E01'
PASS decode('TIS-620', '%DB') is 'U+FFFD'
PASS decode('ISO-8859-11', '%80') is 'U+20AC'
PASS decode('ISO-8859-11', '%96') is 'U+2013'
PASS decode('ISO-8859-11', '%A0') is 'U+00A0'
PASS decode('ISO-8859-11', '%A1') is 'U+0E01'
PASS decode('ISO-8859-11', '%DB') is 'U+FFFD'
PASS decode('windows-874', '%80') is 'U+20AC'
PASS decode('windows-874', '%96') is 'U+2013'
PASS decode('windows-874', '%A0') is 'U+00A0'
PASS decode('windows-874', '%A1') is 'U+0E01'
PASS decode('windows-874', '%DB') is 'U+FFFD'
PASS decode('dos-874', '%80') is 'U+20AC'
PASS decode('dos-874', '%96') is 'U+2013'
PASS decode('dos-874', '%A0') is 'U+00A0'
PASS decode('dos-874', '%A1') is 'U+0E01'
PASS decode('dos-874', '%DB') is 'U+FFFD'
PASS decode('UTF-7', '+AD4') is 'U+002B/U+0041/U+0044/U+0034'
PASS decode('utf-7', '+AD4') is 'U+002B/U+0041/U+0044/U+0034'
PASS decode('UTF-16LE', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('unicodeFEFF', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('UTF-16', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('ISO-10646-UCS-2', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('UCS-2', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('Unicode', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('csUnicode', '%69%D8%D6%DE') is 'U+D869/U+DED6'
PASS decode('UTF-16BE', '%D8%69%DE%D6') is 'U+D869/U+DED6'
PASS decode('unicodeFFFE', '%D8%69%DE%D6') is 'U+D869/U+DED6'
PASS decode('csiso2022kr', '%41%42%43%61%62%63%31%32%33%A0') is 'U+FFFD'
PASS decode('hz-gb-2312', '%41%42%43%61%62%63%31%32%33%A0') is 'U+FFFD'
PASS decode('iso-2022-cn', '%41%42%43%61%62%63%31%32%33%A0') is 'U+FFFD'
PASS decode('iso-2022-cn-ext', '%41%42%43%61%62%63%31%32%33%A0') is 'U+FFFD'
PASS decode('iso-2022-kr', '%41%42%43%61%62%63%31%32%33%A0') is 'U+FFFD'
PASS successfullyParsed is true

TEST COMPLETE
